Clean up false-positive error logs for MS SQL CDC
Without this change logs have thousands of false positive messages that pipeline is exausted. The issue is that "__$sdc.txn_window" is not available in MS SQL and should not be attempted. Ideally solution will be to check column presence in target table with TableContext and skip execution if not founded.